eliminar y modificar meetings y users

+++ app/Http/Controllers/RegitrationController.php
@@ -1,7 +1,8 @@
 <?php
 
 namespace App\Http\Controllers;
-
+use App\Models\Meeting;
+use App\Models\User;
 use Illuminate\Http\Request;
 
 class RegitrationController extends Controller
@@ -23,26 +24,28 @@ class RegitrationController extends Controller
         $meeting_id = $request->input('meeting_id');
         $user_id = $request->input('user_id');
 
-        $meeting = [
-            'title' => 'Title',
-            'description' => 'Description',
-            'time' => 'Time',
-            'view_meeting' => [
-                'href' => 'api/v1/meeting/1',
-                'method' => 'GET'
+        $meeting = Meeting::findOrFail($meeting_id);
+        $user = User::findOrFail($user_id);
+
+        $message = [
+            'msg' => 'User is already registered for meeting',
+            'meeting' => $meeting,
+            'user' => $user,
+            'unregistered' => [
+                'href' => 'api/v1/regitration/'. $meeting_id,
+                'method' => 'DELETE'
             ]
         ];
-
-        $user = [
-            'name' => 'Name'
-        ];
-
+        if($meeting->users()->where('users.id', $user_id)->first()){
+            return response()->json($message, 404);
+        };
+        $user->meetings()->attach($meeting);
         $response = [
             'msg' => 'User registered for meeting',
             'meeting' => $meeting,
             'user' => $user,
             'unregistered' => [
-                'href' => 'api/v1/registration/1',
+                'href' => 'api/v1/regitration/'. $meeting_id,
                 'method' => 'DELETE'
             ]
         ];
@@ -58,26 +61,15 @@ class RegitrationController extends Controller
      */
     public function destroy($id)
     {
-        $meeting = [
-            'title' => 'Title',
-            'description' => 'Description',
-            'time' => 'Time',
-            'view_meeting' => [
-                'href' => 'api/v1/meeting/1',
-                'method'=> 'GET'
-            ]
-        ];
-
-        $user =[
-            'name'=> 'Name'
-        ];
+        $meeting = Meeting::findOrFail($id);
+        $meeting->users()->detach();
 
         $response =[
-            'msg' => 'Usr unregistered for meeting',
+            'msg' => 'User unregistered for meeting',
             'meeting' => $meeting,
-            'user' => $user,
+            'user' => 'tbdd',
             'registered' =>[
-                'href' => 'api/v1/registration',
+                'href' => 'api/v1/meeting/regitration',
                 'method' => 'POST',
                 'params' => 'user_id, meeting_id'
             ]
